Initial commit
This is a prototype to see if these changes would work. Please note
that these are incomplete and should not be pushed to any production
server.

Still need to adjust layout for kiosk display, remove Send To menu, and
confirm users will not be able to log in to any accounts using this
view.